+++ public/app/routes/routes.tsx
@@ -32,14 +32,6 @@ export function getAppRoutes(): RouteDescriptor[] {
     // In order to make it possible we need to register them first due to how `<Switch>` is evaluating routes. (This will be unnecessary once/when we upgrade to React Router v6 and start using `<Routes>` instead.)
     ...getAppPluginRoutes(),
     {
-      path: '/finder',
-      pageClass: 'page-folders',
-      routeName: FolderRoutes.Browse,
-      component: SafeDynamicImport(
-        () => import(/* webpackChunkName: "FoldersPage" */ '../features/browse-resources/BrowseFoldersPage')
-      ),
-    },
-    {
       path: '/',
       pageClass: 'page-dashboard',
       routeName: DashboardRoutes.Home,
@@ -84,7 +76,7 @@ export function getAppRoutes(): RouteDescriptor[] {
       pageClass: 'page-folders',
       routeName: FolderRoutes.Browse,
       component: SafeDynamicImport(
-        () => import(/* webpackChunkName: "FoldersPage" */ 'app/features/browse-resources/BrowseFoldersPage')
+        () => import(/* webpackChunkName: "FoldersPage" */ 'app/features/browse-resources/BrowseResourcesPage')
       ),
     },
     {
